% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_inflow.R
\name{add_inflow}
\alias{add_inflow}
\title{Add inflow with size proportional circles to the map}
\usage{
add_inflow(filname, varName, sizeMax = NULL)
}
\arguments{
\item{filname}{The file (including path) containing the inflow data table}

\item{varName}{The variable used to plot as defined as column name in the
table.}

\item{sizeMax}{The maximum value of the parameter to scale the point size.
If NULL it the maximum value is used, but it can be better to decrease the
sizeMax value in order to visualize small inflows better.}
}
\description{
Add inflow with size proportional circles to the map
}
