% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to the plot}
\usage{
add_logo(
  logo_filename,
  position = "bottomright",
  size = 1,
  indent = 0,
  bg_col = NULL
)
}
\arguments{
\item{logo_filename}{Filename including .png ending}

\item{position}{One of "bottomright", "bottomleft", "bottom", "right", "left",
"center", "topright", "topleft" or "top".}

\item{size}{Numeric value defining the logo size. It is a factor applied to
1/10 of the x-axis length}

\item{indent}{Numeric value (usually between 0 and 1) defining the indent from
the plot frame. like for the size this is a factor refered to 1/10 of x or y
dimension of the plot}

\item{bg_col}{The background color of the logo (NULL does not draw any
background). This can be an R-Color as character string or entered as
\code{\link[=rgb]{rgb()}}, so transperancy can be included.}
}
\value{
Logo is added to the plot in the active device
}
\description{
The logo needs to be in the package extdata/logo folder as png file
}
\examples{
# empty plot
plot(NA, xlim = c(0, 2), ylim = c(0, 4),
  type = "n", xaxt = "n", yaxt = "n", xlab = "", ylab = ""
)
# add logo
add_logo(
  logo_filename = "KWB_Logo.png",
  position = "bottomright",
  size = 2,
  indent = 0.1,
  bg_col = rgb(red = 0, green = 0, blue = 0, alpha = 0.5)
)

}
