% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_polygons.R
\name{add_polygons}
\alias{add_polygons}
\title{Add polygons to a map}
\usage{
add_polygons()
}
\description{
Polygons are drawn in different types of gray and names are added, without
overlapping the catchment boundaries or any river
}
