% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_qsimVis_id.R
\name{add_qsimVis_id}
\alias{add_qsimVis_id}
\title{Adds the river ID of qsimVis package to the table}
\usage{
add_qsimVis_id(aggregated_data, translation_table)
}
\arguments{
\item{aggregated_data}{An input dataframe. Either manually entered or created
by one of \code{\link[=deviating_hours]{deviating_hours()}},\code{\link[=adverse_deviation_from_reference]{adverse_deviation_from_reference()}} or
\code{\link[=critical_events]{critical_events()}}}

\item{translation_table}{Project specific translation dataframe that
contains verknet data and project specific IDs. Column names must be
\itemize{
\item "ID": Verknet ID or ID of manually added rivers which are also the
filnames of the river csv tables
\item "verknet_BWaStrIdNr": The water body number in verknet. If NA the river
information is assumed to be added manually.
\item "Bezeichnung": The real water body name
\item "section_name": The project specific ID of the whole river or a river
section. Can be multiple IDs per waterbody, separated by ",", however, must
be unique.
}}
}
\value{
A data extended aggregated_data data frame by qsimVis_river and qsimVis_source
}
\description{
River ID can either be defined by verknet data (see details) or a manually
entered. The output tables lists the the source of river information and
the defined size of the river for plotting as additional columns.
}
\details{
The Verknet data is available here: https://www.gdws.wsv.bund.de/DE/service/karten/03_VerkNet-BWaStr/VerkNet-BWaStr_node.html
}
