% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_river_legend.R
\name{add_river_legend}
\alias{add_river_legend}
\title{Add a legend for the river color to the plot}
\usage{
add_river_legend(ext_rivers, LegendTitle = "", LegendLocation = "right", ...)
}
\arguments{
\item{ext_rivers}{A list of rivers loaded with \code{\link[=prepare_rivers]{prepare_rivers()}} and extended
with \code{\link[=value_to_classes]{value_to_classes()}}}

\item{LegendTitle}{String with title of the legend}

\item{LegendLocation}{Either "top" or "right" (outside the plot margin)}

\item{...}{Additional legend layout arguments}
}
\description{
Add a legend for the river color to the plot
}
