% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_info_qsim_ID.R
\name{add_site_info}
\alias{add_site_info}
\title{Adds site info based on qsimd Ids to a table}
\usage{
add_site_info(df_in, v_qsim_ids)
}
\arguments{
\item{df_in}{A Dataframe containing site specific data in rows}

\item{v_qsim_ids}{A vector of qsim ids with the same length as the number of
rows of df_in.}
}
\value{
The dataframe df_in extented 4 columns that contain site information
}
\description{
Adds site info based on qsimd Ids to a table
}
