% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_deviation.R
\name{adverse_deviation_from_reference}
\alias{adverse_deviation_from_reference}
\title{Relative adverse deviation from a reference}
\usage{
adverse_deviation_from_reference(dataFrame, reference, worst = 0)
}
\arguments{
\item{dataFrame}{Data frame with time in first column and parameter
values per site. If no reference data frame is provided it must include
the reference_vector as column.}

\item{reference}{Column name of the reference in dataFrame or a seperate
data frame that contains the same column names as dataFrame}

\item{worst}{A numerical value which is the worst case}
}
\value{
Numeric value between 0 and 1
}
\description{
The cumulative sum of all adverse deviations.
}
\details{
First the similarity the data vector and the reference vector is calculated.
Only complete pairs (no NA values) are used. For each data pair the
quotient between data and reference is calcutaled. If data > reference the
value is set to 1. All quotients  are cumulated (-> absolute similarity).
This can be maximum the number of data pairs. When deviding by the number of
data pairs, the relative similarity is obtained. One minus the relative
similarity is the negative deviation.
}
