% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_rivers}
\alias{available_rivers}
\title{Lists all filenames of csv files in a folder}
\usage{
available_rivers(path)
}
\arguments{
\item{path}{The file path}
}
\description{
Lists all filenames of csv files in a folder
}
