% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin.R
\name{Berlin_add_boarder}
\alias{Berlin_add_boarder}
\title{Add Berlin boarder to a map}
\usage{
Berlin_add_boarder(bg_color = "gray60", frame = "black")
}
\arguments{
\item{bg_color}{Character string or \code{\link[=rgb]{rgb()}} for the polygon background color}

\item{frame}{The color of the polygon frame. If NA, no frame will be drawn.}
}
\description{
Polygons is drawn in lightgray
}
