% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin.R
\name{Berlin_add_catchments}
\alias{Berlin_add_catchments}
\title{Add Berlin combined sewer catchments to a map}
\usage{
Berlin_add_catchments(
  plot_names = FALSE,
  highlight_catchments = NULL,
  highlight_style = "shaded"
)
}
\arguments{
\item{plot_names}{If TRUE, names of the districts will be included in the plot}

\item{highlight_catchments}{Character vector containing the catchments names
of catchments to be highlighted}

\item{highlight_style}{Either "shaded" (default) or a specific color, which
cab also be defined by \code{\link[=rgb]{rgb()}}}
}
\description{
Polygons are drawn in different types of gray
}
