% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin.R
\name{Berlin_add_poi}
\alias{Berlin_add_poi}
\title{Add Points of interest in Berlin}
\usage{
Berlin_add_poi(
  poiType,
  poiTitle = "",
  sw_connection = FALSE,
  plotNames = FALSE,
  lineColor = "white",
  fillColor = "brown",
  pch = 21,
  pCex = 1.5,
  dashed_connection = TRUE,
  legendPosition = "topleft",
  rivers = NULL
)
}
\arguments{
\item{poiType}{One of the types defined in the poi-Table ("wwtp", "dwtp", or
further manually added types)}

\item{poiTitle}{A character describing the POI type, which is used as legend
text.}

\item{sw_connection}{If TRUE an arrow is drawn from the POI to i) the point
on the map defined by link_longitued and link_latitude or ii) a specified
outlet into or inlet from a water body. If both locations are defined, the
second one is used.}

\item{plotNames}{If TRUE, names of the POI will be plotted next to the points}

\item{lineColor, fillColor}{Character string or \code{\link[=rgb]{rgb()}} for the frame and fill
color of the POI, respectively}

\item{pch, pCex}{The point type (see pch in \code{\link[=plot]{plot()}}) and size (numeric value)
of the  POI}

\item{dashed_connection}{If TRU (and sw_connection is TRUE) the connection
line will be a dashed line of both colors, lineColor and fillColor}

\item{legendPosition}{Position of the legend ("topleft" or "topright").
If NULL, no POI legend will be plotted.}

\item{rivers}{A list of rivers loaded by \code{\link[=prepare_rivers]{prepare_rivers()}}. Only needed if
connections to surface waters are drawn and defined by river parameters, such
as ID and river kilometer.}
}
\description{
Polygons is drawn in lightgray
}
