% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin.R
\name{Berlin_add_waterbodies}
\alias{Berlin_add_waterbodies}
\title{Add Berlin boarder to a map}
\usage{
Berlin_add_waterbodies(bg_color = "lightblue")
}
\arguments{
\item{bg_color}{Character string or \code{\link[=rgb]{rgb()}} for the polygon background color}
}
\description{
Polygons is drawn in lightgray
}
