% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classes_to_color}
\alias{classes_to_color}
\title{Returns the days of time shift for all years part of a date vector}
\usage{
classes_to_color(class_levels, colorVector = NULL)
}
\arguments{
\item{class_levels}{Levels of a class vector in correct order}

\item{colorVector}{A vector of colors as HEX. By default, the MisaColor will
be used}
}
\value{
A list of dates of timeshifts per year
}
\description{
Returns the days of time shift for all years part of a date vector
}
