% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_events.R
\name{critical_events}
\alias{critical_events}
\title{Count Events of deficits}
\usage{
critical_events(
  dataFrame,
  separating_hours,
  deficiency_hours,
  threshold,
  recovery_value = NULL,
  return_event_positions = FALSE
)
}
\arguments{
\item{dataFrame}{Data frame with time in first column and parameter
values per site.}

\item{separating_hours}{Minimal time period between two values below
threshold to count as separate events.}

\item{deficiency_hours}{Minimal time period of values below threshold.}

\item{threshold}{Numeric in the same unit as the data vector}

\item{recovery_value}{If this is not NULL (default) two events are only
separated if a numerical revocvery value is exceeded between two deficits.}

\item{return_event_positions}{Instead the number of events, the events
starting and end positions are returned, corresponding to the data vector}
}
\value{
Either a number of events or a data frame with event start and end
position
}
\description{
Counts the Number of intervals where a minimum to time in a row are below
a defined threshold value. Events are separated by a specified period of
time above that threshold. Furthermore, the exceedance
of a value can also be a separation criterion.
}
\details{
If deficiency_hours and seperating_hours lead to decimal number of data
points, the numerical is rounded to the next integer. For example:
a minimum deficiency time of 20 minutes in a 15 minute resolution would mean
at least 2.3 data points must be below threshold values which is rounded
to 2 data points, while a deficiency time of 25 minutes would result in 3
data points that needed to be below the threshold.
}
