% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daysOfTimeShift}
\alias{daysOfTimeShift}
\title{Returns the days of time shift for all years part of a date vector}
\usage{
daysOfTimeShift(all_dates)
}
\arguments{
\item{all_dates}{Date-Time vector as character}
}
\value{
A list of dates of timeshifts per year
}
\description{
Returns the days of time shift for all years part of a date vector
}
