% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deficiency_hours.R
\name{deficiency_hours}
\alias{deficiency_hours}
\title{Hours of defficienc within one event}
\usage{
deficiency_hours(dataFrame, thresholds = c(0.5, 1, 1.5, 2, 5))
}
\arguments{
\item{dataFrame}{Data frame with time column "posicDateTime and parameter
values per site}

\item{thresholds}{Oxygen threshold values used for the assessment in mg/L}
}
\value{
Data frane with rows per site and columns per threshold
}
\description{
Hours of defficienc within one event
}
