% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_parameter_data.R
\name{filter_parameter_data}
\alias{filter_parameter_data}
\title{Filters the parameter data frame by time and sites}
\usage{
filter_parameter_data(
  dataFrame,
  sites = "",
  tBeg = min(dataFrame$posixDateTime, na.rm = TRUE),
  tEnd = max(dataFrame$posixDateTime, na.rm = TRUE)
)
}
\arguments{
\item{dataFrame}{Data frame loaded by a MiSa function (see details)}

\item{sites}{Vector of strings containing the names of considered sites,
as column names of the data frame. If "" (default) all sites are included}

\item{tBeg}{POSIX-Value with a start time of the observeration interval}

\item{tEnd}{POSIX-Value with an end time of the observeration interval}
}
\value{
A filtered data frame with the same columns as the input data frame
}
\description{
Filters the loaded data frame by sites and time
}
\details{
The name of the site column must be "site", the name of the timestamp column
must be "posixDateTime".
}
