% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findDateFormat}
\alias{findDateFormat}
\title{Takes a date as character and returns the most likely format if possible}
\usage{
findDateFormat(chracterDates, return_positions = FALSE, check_format = FALSE)
}
\arguments{
\item{chracterDates}{A vector of dates as character}

\item{return_positions}{If TRUE, a dataframe with start and stop positions
of the part of the datetime is returned, which can be used to cut strings
at certain position. If FALSE (default) the most probably date format is
returned which can be used as format argument in POSIX functions}

\item{check_format}{If TRUE, all parts of the timestampt are checked against
their constraints (ex. days must have values between 1 and 31)}
}
\description{
Takes a date as character and returns the most likely format if possible
}
