% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{flow_weighted_mean}
\alias{flow_weighted_mean}
\title{Flow weighted mean per site}
\usage{
flow_weighted_mean(dataFrame, df_flow)
}
\arguments{
\item{dataFrame}{Data frame with time column "posixDateTime" and parameter
values per site. NA values are not removed, make sure there are no NA values
in the time series.}

\item{df_flow}{Data frame in the same format as dataFrame with flow values
per site.}
}
\value{
Data frame with rows per site and columns flow proportional mean
}
\description{
Flow weighted mean per site
}
