% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_multipleNA.R
\name{insert_downstreamNA}
\alias{insert_downstreamNA}
\title{Substitution of one or more missing downstream values}
\usage{
insert_downstreamNA(data_vector)
}
\arguments{
\item{data_vector}{Numeric vector of measurements (including NA values).
Values must correspond to ascending river length (from downstream to
upstream site)}
}
\value{
Data vector with substituted NA values
}
\description{
Substitution of one or more missing downstream values
}
