% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_events.R
\name{load_events}
\alias{load_events}
\title{Loads critical rain events for scenario assessment}
\usage{
load_events(filename)
}
\arguments{
\item{filename}{file name (including path) of the event definition table}
}
\value{
Table as defined in 'extdata/misa_data' filtered
}
\description{
Loads critical rain events for scenario assessment
}
