% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_polygons.R
\name{load_geo}
\alias{load_geo}
\title{Load a package geo file}
\usage{
load_geo(region, Rdata_file)
}
\arguments{
\item{region}{The region of Geo data as defined in the package extdata folder}

\item{Rdata_file}{The name of the file (without ".RData")}
}
\description{
in the package geo files are stored as Rdata with three objects
gis_coordinates (geomitries), gis_data (everything except for geometries)
and readme (list of further information)
}
