% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rivers.R
\name{load_rivers}
\alias{load_rivers}
\title{Reads river data}
\usage{
load_rivers(qsimVis_IDs, path_manual = NULL)
}
\arguments{
\item{qsimVis_IDs}{A character vector of qsimVis IDs as listed in the
translation table}

\item{path_manual}{A path to manually added river tables. If NULL (default)
no manually added rivers are assumed.}
}
\description{
Reads all required verknet river csv files, optionally complemented by
manually added river files. Each csv defines a river stretch and must contain
4 columns:
x: longitude
y: lattitude
km: distance from the river estuary
distance_to_neighbour: distance to the previous point of the river in km
}
