% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timeframes.R
\name{load_timeframes}
\alias{load_timeframes}
\title{Loads a data table containing time frames for events of interest}
\usage{
load_timeframes(filename)
}
\arguments{
\item{filename}{file name (including path) of the event definition table}
}
\value{
Table as defined in 'extdata/misa_data' filtered
}
\description{
Loads a data table containing time frames for events of interest
}
