% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin.R
\name{nTypes}
\alias{nTypes}
\title{Counter function instance}
\usage{
nTypes(reset = FALSE)
}
\arguments{
\item{reset}{If \code{TRUE}, the internal counter is reset to 0. If \code{FALSE} (default),}
}
\value{
A function with an internal counter.
}
\description{
A function created by \code{poiType_counter()} that keeps track of the number of
added points.
}
