% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_empty_map.R
\name{plot_empty_map}
\alias{plot_empty_map}
\title{Creates an map Plot, dimensions based on rivers and Qsim output data}
\usage{
plot_empty_map(bbox = NULL, rivers = NULL, plot_toner = FALSE)
}
\arguments{
\item{bbox}{List as Bounding box with two vectors of length 2 for
c(x_min, x_max) and c(y_min, y_max). If NULL, rivers will be used to
define the bounding box. bbox overwrites rivers for bounding box definition.}

\item{rivers}{List of data frames containing a river stretch (longitude and
latitude) and values to be plotted. Created by functions \code{\link[=load_rivers]{load_rivers()}} and
\code{\link[=extend_riverTable]{extend_riverTable()}}}

\item{plot_toner}{If TRUE a toner map is used as background}
}
\description{
Creates an map Plot, dimensions based on rivers and Qsim output data
}
