% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin.R
\name{poiType_counter}
\alias{poiType_counter}
\title{Create a counter for the number of added points}
\usage{
poiType_counter()
}
\value{
A function that increments or resets an internal counter.
}
\description{
Creates a closure that keeps track of how many times it has been incremented.
Calling the returned function without \code{reset = TRUE} increases the counter by
one and returns the updated value. If \code{reset = TRUE}, the counter is reset to
zero.
}
\examples{
counter <- qsimVis:::poiType_counter()
counter()      # 1
counter()      # 2
counter(TRUE)  # 0

}
