% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rivers.R
\name{prepare_rivers}
\alias{prepare_rivers}
\title{Load river stretches, merge sites, add simulation result and fill gaps}
\usage{
prepare_rivers(
  mapping_table,
  aggregated_data,
  value_column,
  path_manual = NULL,
  gap_filling = "steps"
)
}
\arguments{
\item{mapping_table}{A table that maps the Qsim sites to the verknet and
manually added river names.}

\item{aggregated_data}{A dataframe created by one of \code{\link[=deviating_hours]{deviating_hours()}},
\code{\link[=adverse_deviation_from_reference]{adverse_deviation_from_reference()}} or \code{\link[=critical_events]{critical_events()}} with verknet IDs
as rownames}

\item{value_column}{The column name of the aggregated data table to be
plotted}

\item{path_manual}{A path to manually added river tables. If NULL (default)
no manually added rivers are assumed.}

\item{gap_filling}{Either "interpolation" or "steps".
Defines if NA values between two locations are either interpolated (default)
or kept constant based on an upstream value.}
}
\value{
A list of river stretches, including values from the simulation
}
\description{
Load river stretches, merge sites, add simulation result and fill gaps
}
