% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim_prepare.R
\name{QSIM_prepare_multiple}
\alias{QSIM_prepare_multiple}
\title{Read Qsim output from multiple sub-files of one model}
\usage{
QSIM_prepare_multiple(
  qsim_output_file_list,
  parameter_name = "VO2",
  date_column_name = "date",
  id_column_name = "ID",
  km_column_name = "km",
  flow_column_name = "Q",
  section_column_name = NULL
)
}
\arguments{
\item{qsim_output_file_list}{A list of Qsom output files (including path).
The format of the output files need to be  a csv, ";" - seperated
and "," as a decimal separator (German Style).}

\item{parameter_name}{A String defining the column name of the parameter to
be visualized. "VO2" is the default (Oxygen concentration).}

\item{date_column_name, id_column_name, km_column_name}{A String defining the
obligatory column names.}

\item{flow_column_name, section_column_name}{String defining the column names.
If NULL, the columns will be filled by NA values (flow column) or empty
chracaters (section column)}
}
\value{
A list of 2 tables, where the number of rows correspond to the number if
timestamps and the number of columns corresponds to the number of sites plus
1 timestamp column. Table 1 contains information about the defined parameter.
Table 2 contains flow information if available.

Both tables are sorted by the timestamp which are in ETC/GMT+1 (Winter time
in central Europe)
}
\description{
This function executes the function \code{\link[=QSIM_prepare]{QSIM_prepare()}} multiple times and
combines the tables.
}
