% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rivers.R
\name{qsim_to_verknet_id}
\alias{qsim_to_verknet_id}
\title{Changes Qsim Ids to Verknet IDs}
\usage{
qsim_to_verknet_id(aggregated_data, translation_table)
}
\arguments{
\item{aggregated_data}{A dataframe created by one of \code{\link[=deviating_hours]{deviating_hours()}},
\code{\link[=adverse_deviation_from_reference]{adverse_deviation_from_reference()}} or \code{\link[=critical_events]{critical_events()}}}

\item{translation_table}{Data frame of at least two columns: column number 1:
Verknet ID, column with name "sim_ID": Qsim ID}
}
\description{
Changes Qsim Ids to Verknet IDs
}
