% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_excel.R
\name{save_as_excel}
\alias{save_as_excel}
\title{Writes aggregated data of one period of time into Excel file}
\usage{
save_as_excel(list_of_aggregated_data, path)
}
\arguments{
\item{list_of_aggregated_data}{A named list of data frames}

\item{path}{Path and filename}
}
\description{
Each data frame of the list is converted into an excel sheet
}
