% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveActiveDevice.R
\name{saveActiveDevice}
\alias{saveActiveDevice}
\title{Saves the graphic in the active device either as svg or png}
\usage{
saveActiveDevice(filename, path = NULL, type = "vector", resolution = "medium")
}
\arguments{
\item{filename}{The Name of the file without suffix (ex. "plot")}

\item{path}{Path where to store the file. If no path is defined, the current
working directory is used.}

\item{type}{The default "vector" saves the graphic as svg, everything else
saves as png (ex. "")}

\item{resolution}{One of "low", "medium" or "high" (only used for png files)}
}
\description{
Saves the graphic in the active device either as svg or png
}
