% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_info_qsim_ID.R
\name{site_info_from_qsimID}
\alias{site_info_from_qsimID}
\title{Splits the Qsim ID into 4 differnt informations}
\usage{
site_info_from_qsimID(qsim_id)
}
\arguments{
\item{qsim_id}{A character that is a combination of
river, section_id, section_name and kilometer}
}
\value{
A list of 4, including river name, section id, section name and km
}
\description{
Splits the Qsim ID into 4 differnt informations
}
\details{
The format of the qsim_id must be "RRR_SSSS.rrrr_xx.x"
where RRR is a character string for the river name, SSSS is a character for
the section id, rrrr is a character string for section name and  xx.x is a
numeric value of kilometers. The length of character strings may differ.
}
