% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{General statistics per site}
\usage{
stats(dataFrame)
}
\arguments{
\item{dataFrame}{Data frame with time column "posixDateTime" and parameter
values per site. NA values are not removed, make sure there are no NA values
in the time series.}
}
\value{
Data frame with rows per site and columns for minimum, median,
mean, standard diviation and maximum
}
\description{
General statistics per site
}
