% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_output}
\alias{view_output}
\title{View the output table with rounded results}
\usage{
view_output(ot, digits)
}
\arguments{
\item{ot}{The output table created with one of the aggregating functions}

\item{digits}{The number of digits}
}
\description{
View the output table with rounded results
}
