% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coloured_rivers.R
\name{add_coloredRivers}
\alias{add_coloredRivers}
\title{Adds color scaled river data to a map}
\usage{
add_coloredRivers(ext_rivers)
}
\arguments{
\item{ext_rivers}{A list of rivers loaded with \code{\link[=prepare_rivers]{prepare_rivers()}} and extended
with \code{\link[=value_to_classes]{value_to_classes()}}}
}
\description{
Adds color scaled river data to a map
}
