% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coloured_rivers.R
\name{add_coloredRivers}
\alias{add_coloredRivers}
\title{Adds colour scaled riverdata to a map}
\usage{
add_coloredRivers(
  rivers,
  output_table,
  sixBreaks,
  varName,
  dataType = "time",
  LegendTitle = dataType,
  LegendLocation = "right"
)
}
\arguments{
\item{rivers}{A list of rivers loaded with \code{\link[=load_rivers]{load_rivers()}}}

\item{output_table}{Table of aggregated data created by one of the
functions \code{\link[=critical_events]{critical_events()}}, \code{\link[=adverse_deviation_from_reference]{adverse_deviation_from_reference()}} or
\code{\link[=critical_events]{critical_events()}}}

\item{sixBreaks}{Breaks defining the lower limits of the categories.}

\item{varName}{The column name of the agregated data in the output_table}

\item{dataType}{Used only for the Legend. If "time" is used, it is assumed
that the first water quality category is between 0 and a value above 0, while
for the counting of "events" the first category is 0 only.}

\item{LegendTitle}{String with title of the legend}

\item{LegendLocation}{Either "top" or "right" (outside the plot margin)}
}
\description{
Adds colour scaled riverdata to a map
}
