% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_polygons.R
\name{Berlin_add_boarder}
\alias{Berlin_add_boarder}
\title{Add Berlin boarder to a map}
\usage{
Berlin_add_boarder(bg_color = "gray60")
}
\arguments{
\item{bg_color}{Character string or \code{\link[=rgb]{rgb()}} for the polygon background color}
}
\description{
Polygons is drawn in lightgray
}
