% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviating_hours.R
\name{deviating_hours}
\alias{deviating_hours}
\title{Hours of defficienc within one event}
\usage{
deviating_hours(dataFrame, thresholds = c(0.5, 1, 1.5, 2, 5), dev_type = "elt")
}
\arguments{
\item{dataFrame}{Data frame with time column "posicDateTime and parameter
values per site}

\item{thresholds}{Oxygen threshold values used for the assessment in mg/L}

\item{dev_type}{String definint the type of deviation. "elt" for equal or
lower than, "egt" for equal or greater than.}
}
\value{
Data frane with rows per site and columns per threshold
}
\description{
Hours of defficienc within one event
}
