% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_riverTable}
\alias{extend_riverTable}
\title{Adds quality categories to the river table}
\usage{
extend_riverTable(river_table, aggregated_data, varName, sixBreaks)
}
\arguments{
\item{river_table}{Table of one of the rivers loaded with
\code{\link[=load_rivers]{load_rivers()}}}

\item{aggregated_data}{A dataframe created by one of \code{\link[=deviating_hours]{deviating_hours()}},
\code{\link[=adverse_deviation_from_reference]{adverse_deviation_from_reference()}} or \code{\link[=critical_events]{critical_events()}}}

\item{varName}{The column name of the agregated data in the output_table}

\item{sixBreaks}{Breaks defining the lower limits of the categories.}
}
\value{
The input river_table extended by three columns: 1) "value" containing the
exact value, 2) "quality" containing 6 different quality categories based on
the defined breaks, 3) "color" assigning th color for plotting
}
\description{
As a result, each location is linked to a water quality value.
}
\details{
The qsim_misa_table does not provide information for every location within
the river, since the local resolution of the qsim output differs from the
river course data. Values between two known river sites are interpolated.
The color scale cannot be changed. Since in all misa assessment parameters
low values refer to a good water quality, the colors range from green for low
values to red for high values.
}
