% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_empty_map.R
\name{plot_empty_map}
\alias{plot_empty_map}
\title{Creates an map Plot, dimensions based on rivers and Qsim output data}
\usage{
plot_empty_map(rivers, plot_toner = FALSE)
}
\arguments{
\item{rivers}{List of data frames containing a river stretch (longitude and
latitude sorted in direction of flow) combined with the Qsim site ID}

\item{plot_toner}{If TRUE a toner map is used as background}
}
\description{
Creates an map Plot, dimensions based on rivers and Qsim output data
}
