% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_qsim.R
\name{QSIM_prepare}
\alias{QSIM_prepare}
\title{Read Qsim output and save or return table for aggregation and visualization}
\usage{
QSIM_prepare(qsim_output_file, parameter_name = "VO2")
}
\arguments{
\item{qsim_output_file}{filename (including path) in which the output from
Gsim is listed}

\item{parameter_name}{A String defining the column name of the parameter to
be visualized. "VO2" is the default (Oxygen concentration).}
}
\value{
The saved table contains a timestamp column in Central European Time and
several oxygen columns in mg/L
}
\description{
This function combines the Qsim site information of the first 3 columns
to unique sitenames and spreads the Qsim output table into a table with one
time column and several parameter columns (one per site)
}
\details{
Saves table as CSV File ";"-separated if parameter return_table is set to TRUE
}
