% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{site_info_from_qsimID}
\alias{site_info_from_qsimID}
\title{Get River, Section and location information from Qsim IDs}
\usage{
site_info_from_qsimID(assessment_list)
}
\arguments{
\item{assessment_list}{The assessment of a Qsim simulation}
}
\description{
Get River, Section and location information from Qsim IDs
}
