% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_to_classes.R
\name{value_to_classes}
\alias{value_to_classes}
\title{Add two columns class and color}
\usage{
value_to_classes(river_list, classBreaks)
}
\arguments{
\item{river_list}{A list of river tables including a value column created by
\code{\link[=prepare_rivers]{prepare_rivers()}}}

\item{classBreaks}{Breaks defining the limits. The lowerst value is included
in the lowest class. In the remaining classes, the highest value is included.}
}
\value{
The river_dataTable extended by two columns "quality" (class) and
"color"
}
\description{
Add two columns class and color
}
